+++ docs/source/expression_tree/broadcasts.rst
@@ -1,0 +1,11 @@
+Broadcasting Operators
+======================
+
+.. autoclass:: pybamm.TimeBroadcast
+  :members:
+
+.. autoclass:: pybamm.Broadcast
+  :members:
+
+.. autoclass:: pybamm.NumpyBroadcast
+  :members:
