+++ docs/source/expression_tree/index.rst
@@ -7,9 +7,10 @@ Expression Tree
   parameter
   variable
   independent_variable
-  binary_operator
-  unary_operator
   scalar
   matrix
   vector
+  binary_operator
+  unary_operator
   concatenations
+  broadcasts
