+++ docs/source/expression_tree/unary_operator.rst
@@ -19,11 +19,6 @@ Unary Operators
 .. autoclass:: pybamm.Divergence
   :members:
 
-.. autoclass:: pybamm.Broadcast
-  :members:
-
-.. autoclass:: pybamm.NumpyBroadcast
-  :members:
 
 .. autofunction:: pybamm.grad
